Fix a couple of Windows things.
* Remove _WIN32_WINNT check, which was interfering with the libretro
  port. The test can go elsewhere, if it's even still necessary (which I
  suspect it isn't)

* Add SCODE-to-DWORD casts. Hopefully fix warning encountered on gcc

+++ src/shared/c/system_windows.cpp
@@ -120,8 +120,8 @@ const char *GetErrorDescription(DWORD error) {
     if (n == 0) {
         /* Bleargh. */
 
-#define CASE(X) \
-    case (X):   \
+#define CASE(X)      \
+    case (DWORD)(X): \
         return #X
 
         switch (error) {
